requirements.txt
# Core
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
python-dotenv>=1.0.0

# Crawling
crawl4ai>=0.4.0

# Processing
sentence-transformers>=2.6.1
torch>=2.1.0

# Vector DB 
chromadb==0.4.24

# NumPy 
numpy<2.0

# LLM - Google Gemini
google-generativeai>=0.7.0

# API Framework
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Logging and monitoring
python-json-logger>=2.0.0

# Development (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
